refactor: move member model to core domain
