+++ core/domain/src/commonMain/kotlin/com/parksupark/soomjae/core/domain/models/Member.kt
@@ -1,4 +1,4 @@
-package com.parksupark.soomjae.features.post.domain.models
+package com.parksupark.soomjae.core.domain.models
 
 data class Member(
     val id: String,
